NXP-28865: remove useless DBS internal session proxying
